Add wrappers for page.$eval and page.$$eval
These respectively run a passed function against the first
(querySelector) element and against all (querySelectorAll) elements
matching their selector.

The functions accept a function string which is later `eval`-ed to turn it
into a JavaScript function we can pass to puppeteer. This makes it
possible to easily pass dependency-less functions to puppeteer.